search-stats-service/pyproject.toml
[tool.poetry]
name = "search-stats-service"
version = "0.1.0"
description = "Search statistics service for the micro-blog app"
authors = ["Mohammed Mwijaa <mm.mwijaa@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
gunicorn = "^20.0.4"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
requests = "^2.25.1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report html"
testpaths = [
	"tests"
]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
